implemented camera preview
