+++ package.json
@@ -19,8 +19,8 @@
     "axios": "^1.9.0",
     "bcrypt": "^5.1.1",
     "bcryptjs": "^3.0.2",
-    "cloudinary": "^2.7.0",
     "class-variance-authority": "^0.7.1",
+    "cloudinary": "^2.7.0",
     "clsx": "^2.1.1",
     "cookie": "^0.7.0",
     "framer-motion": "^12.18.1",
@@ -34,10 +34,10 @@
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.5.2",
     "react-icons": "^5.5.0",
-    "react-toastify": "^11.0.5",
-    "resend": "^4.6.0",
     "react-konva": "^19.0.7",
     "react-player": "^2.12.0",
+    "react-toastify": "^11.0.5",
+    "resend": "^4.6.0",
     "tailwind-merge": "^3.3.1",
     "zod": "^3.25.62",
     "zustand": "^5.0.5"
